setup and css fixed
