+++ .htaccess
@@ -1,3 +1,3 @@
 RewriteEngine On
-RewriteRule ^([^/]+)/? app/views/application.php?url=$1 [L,QSA]
-RewriteRule .* app/views/application.php?url=$1 [L,QSA]
+RewriteRule ^([^/]+)/? app/application.php?url=$1 [L,QSA]
+RewriteRule .* app/application.php?url=$1 [L,QSA]
