+++ app/controllers/controller.php
@@ -1,8 +1,7 @@
 <?php  
   class Controller extends Database {
-    public static function CreateView($viewName) {
-      require_once("./views/$viewName.php");
-      echo $view_name;
+    public static function CreateView($controllerName, $viewName) {
+      require_once("./views/$controllerName/$viewName.php");
     }
   }
 ?>
