+++ app/controllers/user.php
@@ -1,0 +1,5 @@
+<?php 
+  class User extends Controller {
+
+  }
+?>
