+++ app/stylesheets/components/container.css
@@ -1,5 +1,4 @@
 .container {
-  border: solid red 3px;
   margin-top: 50px;
   text-align: center;
   display: flex;
