+++ app/stylesheets/stylesheets.php
@@ -1,0 +1,11 @@
+<style>
+  <?php 
+    require("stylesheets/application.css");
+
+    // Import CSS Components
+    require("components/container.css");
+
+    // Import CSS pages
+    require("pages/users/index.css");
+  ?>
+</style>
