+++ app/views/users/index.php
@@ -1,0 +1,3 @@
+<div class="container">
+  <h1>Home Page temporary redirect to User Index</h1>
+</div>
