--- stylesheets/pages/home.css
@@ -1,3 +1,0 @@
-.stimulus-title {
-  color: purple;
-}
